Refactor: code review on exchanges services and controllers (using PEPVerification) for improved clarity
