+++ src/controllers/public/v1/data.public.controller.ts
@@ -1,20 +1,12 @@
 import { NextFunction, Request, Response } from 'express';
-import { decryptSignedConsent } from '../../../utils/decryptConsent';
 import { Logger } from '../../../libs/loggers';
-import { validateConsent } from '../../../libs/services/validateConsent';
-import axios from 'axios';
-import { handle } from '../../../libs/loaders/handler';
-import { getCatalogData } from '../../../libs/services/catalog';
-import {
-    pepLeftOperandsVerification,
-    pepVerification,
-} from '../../../utils/pepVerification';
-import { getRepresentation } from '../../../libs/loaders/representationFetcher';
-import { DataExchangeStatusEnum } from '../../../utils/enums/dataExchangeStatusEnum';
-import { processLeftOperands } from '../../../utils/leftOperandProcessor';
-import { DataExchange } from '../../../utils/types/dataExchange';
 import { restfulResponse } from '../../../libs/api/RESTfulResponse';
-import { importDataService } from '../../../services/public/v1/consumer.public.service';
+import {
+    exportDataService,
+    importDataService,
+} from '../../../services/public/v1/data.public.service';
+import { DataExchangeStatusEnum } from '../../../utils/enums/dataExchangeStatusEnum';
+import { DataExchangeResult } from '../../../utils/types/dataExchange';
 
 /**
  * Export data for the provider in the consent flow
@@ -34,96 +26,34 @@ export const exportData = async (
             error: 'Missing params from request payload',
         });
 
-    // Send OK response to requester
-    res.status(200).json({ message: 'OK' });
-
-    // Decrypt signed consent and retrieve token + consentId
-    const decryptedConsent = await decryptSignedConsent(
-        signedConsent,
-        encrypted
-    );
-
-    // Get dataExchange
-    const dataExchange = await DataExchange.findById(
-        decryptedConsent.providerDataExchangeId
-    );
-
     try {
-        // Send validation verification to VisionsTrust to receive user info and DataTypes
-        const validation = await validateConsent(signedConsent, encrypted);
+        // Send OK response to requester
+        res.status(200).json({ message: 'OK' }); // ?
 
-        //eslint-disable-next-line
-        const { verified } = validation;
+        const result: DataExchangeResult = await exportDataService({
+            signedConsent,
+            encrypted,
+        });
 
-        if (!verified) {
-            throw new Error('consent not verified.');
-        }
-
-        for (const dt of decryptedConsent.data) {
-            const { success: pepSuccess } = await pepVerification({
-                targetResource: dt.serviceOffering,
-                referenceURL: decryptedConsent.contract,
+        if (
+            result.exchange !== null &&
+            result.exchange.status === DataExchangeStatusEnum.EXPORT_SUCCESS
+        ) {
+            return restfulResponse(res, 200, { success: true, message: 'OK' });
+        } else {
+            Logger.error({
+                message: result.errorMessage,
             });
-
-            if (pepSuccess) {
-                const [dataResourceSD] = await handle(
-                    getCatalogData(dt.resource)
-                );
-
-                const [data] = await handle(
-                    getRepresentation(
-                        dataResourceSD.representation?.method,
-                        dataResourceSD.representation.url,
-                        dataResourceSD.representation?.credential,
-                        decryptedConsent
-                    )
-                );
-
-                // POST the data to the import service
-                const importResponse = await axios({
-                    url: (decryptedConsent as any).dataConsumer.endpoints
-                        .dataImport,
-                    method: 'POST',
-                    data: {
-                        data: data,
-                        user: (decryptedConsent as any).consumerUserIdentifier
-                            .identifier,
-                        signedConsent: signedConsent,
-                        encrypted,
-                        resource: dt.resource,
-                        apiResponseRepresentation: !!(
-                            dataResourceSD.isPayloadForAPI &&
-                            dataResourceSD.apiResponseRepresentation
-                        ),
-                    },
-                });
-
-                // Process left Operands incrementation
-                if (importResponse?.data?.message === 'OK') {
-                    const names = await pepLeftOperandsVerification({
-                        targetResource: dt.serviceOffering,
-                    });
-                    await processLeftOperands(
-                        names,
-                        decryptedConsent.contract,
-                        dt.serviceOffering
-                    );
-                }
-            } else {
-                await dataExchange.updateStatus(
-                    DataExchangeStatusEnum.PEP_ERROR
-                );
-            }
+            return restfulResponse(res, 400, { success: false, error: '' });
         }
     } catch (err) {
         Logger.error({
             message: err.message,
             location: err.stack,
         });
-        await dataExchange.updateStatus(
-            DataExchangeStatusEnum.CONSENT_EXPORT_ERROR,
-            err.message
-        );
+        return res.status(500).json({
+            error: 'An unexpected error occurred',
+        });
     }
 };
 
@@ -161,7 +91,7 @@ export const importData = async (
             });
         }
 
-        const result = await importDataService({
+        const result: DataExchangeResult = await importDataService({
             data,
             user,
             signedConsent,
@@ -171,11 +101,15 @@ export const importData = async (
             resource,
         });
 
-        if (result.success) {
-            return restfulResponse(res, 200, { message: 'OK' });
+        if (
+            result.exchange !== null &&
+            result.exchange.status === DataExchangeStatusEnum.IMPORT_SUCCESS
+        ) {
+            return restfulResponse(res, 200, { success: true, message: 'OK' });
         } else {
             return restfulResponse(res, 400, {
-                error: result.error,
+                success: false,
+                error: result.errorMessage,
             });
         }
     } catch (error) {
